rest_app/requirements.txt
Django==1.9.8
djangorestframework==3.4.0
psycopg2==2.6.2
gunicorn==19.6.0
coverage==4.1
